Ignore disposition error for stale messages
Messages settled after a link has been closed
trigger exceptions, this is expected. These exceptions
are ignored to avoid creating confusion in application.
Unsettled messages are requeued when a link is closed,
so applications must be prepared to deal with redeliveries.

This can happen for message whose processing spans over
connection recovery.