Retry stream consumer recovery if no local member
A stream can take a few seconds to restart after a node restart, so its
local member may not be available, even though the connection has
recovered.

This commit introduces some retry logic when this happens. Consumers
should use their own session in this, as the broker closes the session
when this errors comes up.