Use work pool to dispatch incoming messages
It is based on the mechanism the RabbitMQ AMQP 091 Java client uses.
The work pool accumulates the message processing tasks and a work
service dispatches the work blocks to an executor.

Each connection creates its own executor service if none is set. An
executor service can be set at the environment level and shared between
all connections. An executor service can be set at the connection level
as well.

References #160