Make Netty kqueue and io_uring transport dependencies optional
Their inclusion can trigger unexpected automatic behavior for some downstream
libraries (gateway webflux server uses io_uring if it is present).

They can be added explictly by a user that wants to use them as the
Netty transport layer in the AMQP client. This is unlikely though,
native epoll and JDK NIO are the defaults and they are usually both fine
for a client application.

Fixes #292